+++ Nickel/Framework/Models/CobaltCoreHandlerResult.cs
@@ -1,4 +1,4 @@
-using System.IO;
+using Nanoray.PluginManager;
 using System.Reflection;
 
 namespace Nickel;
@@ -7,5 +7,5 @@ internal readonly struct CobaltCoreHandlerResult
 {
 	public Assembly GameAssembly { get; init; }
 	public MethodInfo EntryPoint { get; init; }
-	public DirectoryInfo WorkingDirectory { get; init; }
+	public IDirectoryInfo WorkingDirectory { get; init; }
 }